[hotfix] Introduce SlotPoolPendingRequestFailureTest suite
Move testFailingAllocationFailsPendingSlotRequests from SlotPoolImplTest to
SlotPoolPendingRequestFailureTest.
